фикс показа времени кусочка танго
