+++ src/main/java/com/dota2/event/AttackEntity.java
@@ -3,12 +3,14 @@ package com.dota2.event;
 import com.dota2.component.hero.HeroComponent;
 import com.dota2.component.hero.ValuesComponent;
 import com.dota2.effect.CustomEffect;
+import com.dota2.item.Weapon;
 import com.dota2.util.InventoryUtils;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.attribute.EntityAttributeInstance;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.util.hit.EntityHitResult;
@@ -22,6 +24,20 @@ import static com.dota2.component.ModComponents.HERO_COMPONENT;
 import static com.dota2.component.ModComponents.VALUES_COMPONENT;
 
 public class AttackEntity {
+    public static int calculateDamage(PlayerEntity provider) {
+        int totalDamage = 0;
+
+        for (int i = 0; i < provider.getInventory().size(); i++) {
+            ItemStack stack = provider.getInventory().getStack(i);
+
+            if (stack.getItem() instanceof Weapon weapon) {
+                totalDamage += weapon.getDamage();
+            }
+        }
+
+        return totalDamage;
+    }
+
     public static ActionResult onAttackEntity(PlayerEntity playerSource, World world, Hand hand, Entity entity, @Nullable EntityHitResult entityHitResult) {
         if (entity instanceof PlayerEntity playerTarget) {
             HeroComponent heroComponentSource = playerSource.getComponent(HERO_COMPONENT);
@@ -29,7 +45,7 @@ public class AttackEntity {
             if (heroComponentTarget.isHero() && heroComponentSource.isHero() &&
                     !playerSource.isTeammate(playerTarget)) {
                 // Уменьшение хп
-                int damage = InventoryUtils.calculateDamage(playerSource);
+                int damage = calculateDamage(playerSource);
                 EntityAttributeInstance critAttribute = playerSource.getAttributeInstance(CRIT_CHANCE);
                 if (critAttribute != null) {
                     ValuesComponent valuesComponentTarget = playerTarget.getComponent(VALUES_COMPONENT);
